pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "observable-security-agent"
version = "0.1.0"
description = "Multi-cloud security evidence correlation — observability and audit-readiness, not misconfiguration scanning."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "boto3>=1.34",
  "PyYAML>=6.0.1",
  "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
api = ["fastapi>=0.110", "uvicorn[standard]>=0.27", "httpx>=0.27"]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "providers*", "evals*", "instrumentation*", "api*", "fedramp20x*", "agent_loop*", "classification*", "normalization*", "ai*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests that run real subprocesses or otherwise take >1s",
]
